+++ packages/keyring-snap-bridge/src/SnapKeyring.test.ts
@@ -29,6 +29,9 @@ import {
   TrxScope,
   TrxMethod,
   TrxAccountType,
+  XlmScope,
+  XlmMethod,
+  XlmAccountType,
   AccountCreationType,
 } from '@metamask/keyring-api';
 import { SnapManageAccountsMethod } from '@metamask/keyring-snap-sdk';
@@ -216,6 +219,15 @@ describe('SnapKeyring', () => {
     type: TrxAccountType.Eoa,
   };
 
+  const xlmAccount = {
+    id: '890ee179-5ab5-449d-9c25-34e12c1ada68',
+    address: 'GDRF6HX6GXUA74N7LFSXVYPPINW5QRLFPQS4PNFG7HJF6DFQQNT2TI4F',
+    options: {},
+    methods: [...Object.values(XlmMethod)],
+    scopes: [XlmScope.Mainnet, XlmScope.Testnet],
+    type: XlmAccountType.Account,
+  };
+
   // This is an example generic chain, its scope and address are bogus.
   const anyGenericAccount: KeyringAccount = {
     id: '6131c205-26b5-40a8-9626-4761423bb1d5',
@@ -262,6 +274,7 @@ describe('SnapKeyring', () => {
     btcAccountP2tr,
     solDataAccount,
     trxEoaAccount,
+    xlmAccount,
     anyGenericAccount,
   ] as const;
 
@@ -1228,6 +1241,7 @@ describe('SnapKeyring', () => {
           btcAccountP2tr.address,
           solDataAccount.address,
           trxEoaAccount.address,
+          xlmAccount.address,
           anyGenericAccount.address,
         ]);
       });
@@ -1248,6 +1262,7 @@ describe('SnapKeyring', () => {
           btcAccountP2tr.address,
           solDataAccount.address,
           trxEoaAccount.address,
+          xlmAccount.address,
           anyGenericAccount.address,
         ]);
       });
@@ -1460,6 +1475,7 @@ describe('SnapKeyring', () => {
         btcAccountP2tr.address,
         solDataAccount.address,
         trxEoaAccount.address,
+        xlmAccount.address,
         anyGenericAccount.address,
       ]);
     });
@@ -1506,6 +1522,7 @@ describe('SnapKeyring', () => {
           [btcAccountP2tr.id]: { account: btcAccountP2tr, snapId },
           [solDataAccount.id]: { account: solDataAccount, snapId },
           [trxEoaAccount.id]: { account: trxEoaAccount, snapId },
+          [xlmAccount.id]: { account: xlmAccount, snapId },
           [anyGenericAccount.id]: { account: anyGenericAccount, snapId },
         },
       };
@@ -1560,6 +1577,7 @@ describe('SnapKeyring', () => {
       btcTestnetAccount,
       solDataAccount,
       trxEoaAccount,
+      xlmAccount,
     ])('migrates accounts v1: %s', async (expectedAccount: KeyringAccount) => {
       // A v1 account has no scopes, so remove it.
       const state = {
@@ -1583,6 +1601,7 @@ describe('SnapKeyring', () => {
       btcTestnetAccount,
       solDataAccount,
       trxEoaAccount,
+      xlmAccount,
       anyGenericAccount,
     ])(
       'migrates v2 accounts to v1 accounts is noop: %s',
@@ -2243,6 +2262,7 @@ describe('SnapKeyring', () => {
         accounts[8].address,
         accounts[9].address,
         accounts[10].address,
+        accounts[11].address,
       ]);
     });
 
@@ -2261,6 +2281,7 @@ describe('SnapKeyring', () => {
         accounts[8].address,
         accounts[9].address,
         accounts[10].address,
+        accounts[11].address,
       ]);
       expect(console.error).toHaveBeenCalledWith(
         "Account '0xc728514df8a7f9271f4b7a4dd2aa6d2d723d3ee3' may not have been removed from snap 'local:snap.mock':",